Fixed 'sessions' table update during program exit.
Sometimes watta didn't have enough time to update the 'sessions' MySql table with the time of exit, because the window manager killed the application too quickly after closing the main window. With this fix, watta now updates the database within the 'aboutToQuit' message of the MainApplication, which should give it enough time.

Added program name and version to the Main Window title and icon tooltip.